+++ spec/kuber_kit/service_deployer/strategies/docker_spec.rb
@@ -93,12 +93,24 @@ RSpec.describe KuberKit::ServiceDeployer::Strategies::Docker do
 
   it "allows setting env file" do
     expect(subject.docker_commands).to receive(:run).with(
-      shell, "default/auth_app:latest", detached: false, args: ["--name auth_job", "--env-file test.env", "--hostname auth_job", ],
+      shell, "default/auth_app:latest", detached: false, args: ["--name auth_job", "--env-file test.env", "--hostname auth_job"],
       command: "bash", interactive: true, detached: false
     )
 
     service = service_helper.service(:auth_job, attributes: {
       deployer: {image_name: image.name, command_name: "bash", env_file: "test.env"}
+    })
+    subject.deploy(shell, service)
+  end
+
+  it "allows setting env vars" do
+    expect(subject.docker_commands).to receive(:run).with(
+      shell, "default/auth_app:latest", detached: false, args: ["--name auth_job", "--hostname auth_job", "--env RUBY_ENV=development"],
+      command: "bash", interactive: true, detached: false
+    )
+
+    service = service_helper.service(:auth_job, attributes: {
+      deployer: {image_name: image.name, command_name: "bash", env_vars: {RUBY_ENV: "development"}}
     })
     subject.deploy(shell, service)
   end