+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **0.5.9**(to be released)
 - Added an ability to set custom user
+- Allow setting environment variable in docker strategy
 
 **0.5.8**
 - Update gemspec to support ruby 2.5